computer_use_demo/requirements.txt
streamlit>=1.38.0
anthropic[bedrock,vertex]>=0.37.1
jsonschema==4.22.0
boto3>=1.28.57
google-auth<3,>=2
pyobjc-framework-Quartz>=8.0